Use 32-bit integers like the specification states

+++ src/v1_6/messages/clear_charging_profile.rs
@@ -5,13 +5,13 @@ use validator::Validate;
 #[serde(rename_all = "camelCase")]
 pub struct ClearChargingProfileRequest {
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub id: Option<i64>,
+    pub id: Option<i32>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub connector_id: Option<i64>,
+    pub connector_id: Option<i32>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub charging_profile_purpose: Option<ChargingProfilePurposeType>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub stack_level: Option<i64>,
+    pub stack_level: Option<i32>,
 }
 
 #[derive(serde::Serialize, serde::Deserialize, Validate, Debug, Clone, PartialEq, Default)]
