+++ src/v1_6/messages/send_local_list.rs
@@ -5,7 +5,7 @@ use crate::v1_6::types::{AuthorizationData, UpdateStatus, UpdateType};
 #[serde(rename_all = "camelCase")]
 pub struct SendLocalListRequest {
     /// Required. In case of a full update this is the version number of the full list. In case of a differential update it is the version number of the list after the update has been applied.
-    pub list_version: i64,
+    pub list_version: i32,
     /// Optional. In case of a full update this contains the list of values that form the new local authorization list. In case of a differential update it contains the changes to be applied to the local authorization list in the Charge Point. Maximum number of AuthorizationData elements is available in the configuration key: SendLocalListMaxLength
     #[serde(skip_serializing_if = "Option::is_none")]
     pub local_authorization_list: Option<Vec<AuthorizationData>>,
