+++ src/v1_6/types/charging_profile.rs
@@ -7,9 +7,9 @@ use super::{
 #[derive(serde::Serialize, serde::Deserialize, Debug, Clone, PartialEq, Default)]
 #[serde(rename_all = "camelCase")]
 pub struct ChargingProfile {
-    pub charging_profile_id: i64,
+    pub charging_profile_id: i32,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub transaction_id: Option<i64>,
+    pub transaction_id: Option<i32>,
     pub stack_level: u64,
     pub charging_profile_purpose: ChargingProfilePurposeType,
     pub charging_profile_kind: ChargingProfileKindType,
