+++ src/v2_0_1/datatypes/charging_schedule_period_type.rs
@@ -4,13 +4,13 @@
 #[serde(rename_all = "camelCase")]
 pub struct ChargingSchedulePeriodType {
     /// Required. Start of the period, in seconds from the start of schedule. The value of StartPeriod also defines the stop time of the previous period.
-    pub start_period: i64,
+    pub start_period: i32,
     /// Required. Charging rate limit during the schedule period, in the applicable chargingRateUnit, for example in Amperes (A) or Watts (W). Accepts at most one digit fraction (e.g. 8.1).
     pub limit: f64,
     /// Optional. The number of phases that can be used for charging. If a number of phases is needed, numberPhases=3 will be assumed unless another number is given.
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub number_phases: Option<i64>,
+    pub number_phases: Option<i32>,
     /// Optional. Values: 1..3, Used if numberPhases=1 and if the EVSE is capable of switching the phase connected to the EV, i.e. ACPhaseSwitchingSupported is defined and true. It’s not allowed unless both conditions above are true. If both conditions are true, and phaseToUse is omitted, the Charging Station / EVSE will make the selection on its own
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub phase_to_use: Option<i64>,
+    pub phase_to_use: Option<i32>,
 }
