+++ src/v2_0_1/datatypes/cost_type.rs
@@ -5,7 +5,7 @@ use crate::v2_0_1::enumerations::cost_kind_enum_type::CostKindEnumType;
 #[serde(rename_all = "camelCase")]
 pub struct CostType {
     pub cost_kind: CostKindEnumType,
-    pub amount: i64,
+    pub amount: i32,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub amount_multiplier: Option<i8>,
 }
