+++ src/v2_0_1/datatypes/relative_time_interval_type.rs
@@ -2,6 +2,6 @@
 #[derive(serde::Serialize, serde::Deserialize, Debug, Clone, PartialEq, Default)]
 #[serde(rename_all = "camelCase")]
 pub struct RelativeTimeIntervalType {
-    pub start: i64,
-    pub duration: i64,
+    pub start: i32,
+    pub duration: i32,
 }
