+++ src/v2_0_1/messages/reset.rs
@@ -9,7 +9,7 @@ pub struct ResetRequest {
     #[serde(rename = "type")]
     pub request_type: ResetEnumType,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub evse_id: Option<i64>,
+    pub evse_id: Option<i32>,
 }
 
 /// This contains the field definition of the ResetResponse PDU sent by the Charging Station to the CSMS in response to ResetRequest.
